+++ reviewd/src/database/data_source.rs
@@ -1,7 +1,12 @@
-use actix_web::{http, web, HttpResponse};
+use actix_web::{
+    http,
+    web::{Data, Query},
+    HttpResponse,
+};
 use diesel::prelude::*;
 use futures::{future, prelude::*};
 use serde::{Deserialize, Serialize};
+use serde_json::Value;
 
 use super::schema::data_source;
 use crate::database::{build_err_msg, Error, Pool};
@@ -13,13 +18,32 @@ pub(crate) struct DataSourceQuery {
 
 #[derive(Debug, Identifiable, Queryable, Serialize)]
 #[table_name = "data_source"]
-pub(crate) struct DataSourceTable {
-    pub(crate) id: i32,
-    pub(crate) topic_name: String,
-    pub(crate) data_type: String,
+struct DataSourceTable {
+    id: i32,
+    topic_name: String,
+    data_type: String,
 }
 
-pub(crate) fn add_data_source(pool: &web::Data<Pool>, data_source: &str, data_type: &str) -> i32 {
+pub(crate) fn add_data_source_endpoint(
+    pool: Data<Pool>,
+    query: Query<Value>,
+) -> impl Future<Item = HttpResponse, Error = actix_web::Error> {
+    let data_source = query.get("data_source").and_then(Value::as_str);
+    let data_type = query.get("data_type").and_then(Value::as_str);
+
+    let query_result = if let (Some(data_source), Some(data_type)) = (data_source, data_type) {
+        add_data_source(&pool, data_source, data_type)
+    } else {
+        return future::result(Ok(HttpResponse::BadRequest().into()));
+    };
+
+    match query_result {
+        0 => future::result(Ok(HttpResponse::InternalServerError().into())),
+        _ => future::result(Ok(HttpResponse::Created().into())),
+    }
+}
+
+pub(crate) fn add_data_source(pool: &Data<Pool>, data_source: &str, data_type: &str) -> i32 {
     use data_source::dsl;
 
     let new_data_source: Result<DataSourceTable, Error> =
@@ -38,7 +62,7 @@ pub(crate) fn add_data_source(pool: &web::Data<Pool>, data_source: &str, data_ty
     new_data_source.ok().map_or(0, |d| d.id)
 }
 
-pub(crate) fn get_data_source_id(pool: &web::Data<Pool>, data_source: &str) -> Result<i32, Error> {
+pub(crate) fn get_data_source_id(pool: &Data<Pool>, data_source: &str) -> Result<i32, Error> {
     use data_source::dsl;
     pool.get().map_err(Into::into).and_then(|conn| {
         dsl::data_source
@@ -50,7 +74,7 @@ pub(crate) fn get_data_source_id(pool: &web::Data<Pool>, data_source: &str) -> R
 }
 
 pub(crate) fn get_data_source_table(
-    pool: web::Data<Pool>,
+    pool: Data<Pool>,
 ) -> impl Future<Item = HttpResponse, Error = actix_web::Error> {
     let query_result: Result<Vec<DataSourceTable>, Error> =
         pool.get().map_err(Into::into).and_then(|conn| {